added styles in items and renamed
